Moved logic from app.js to Map.js
